+++ actix-http/src/lib.rs
@@ -42,6 +42,7 @@ pub mod h2;
 pub mod header;
 mod helpers;
 mod http_message;
+mod keep_alive;
 mod message;
 mod payload;
 mod requests;
@@ -51,11 +52,12 @@ pub mod test;
 pub mod ws;
 
 pub use self::builder::HttpServiceBuilder;
-pub use self::config::{KeepAlive, ServiceConfig};
+pub use self::config::ServiceConfig;
 pub use self::error::Error;
 pub use self::extensions::Extensions;
 pub use self::header::ContentEncoding;
 pub use self::http_message::HttpMessage;
+pub use self::keep_alive::KeepAlive;
 pub use self::message::ConnectionType;
 pub use self::message::Message;
 #[allow(deprecated)]